+++ R/ggmid_mid_importance.R
@@ -46,9 +46,8 @@
 #'
 ggmid.mid.importance <- function(
     object, type = c("barplot", "dotchart", "heatmap", "boxplot"),
-    theme = NULL, terms = NULL, max.nterms = NULL, ...) {
+    theme = NULL, terms = NULL, max.nterms = 30L, ...) {
   type <- match.arg(type)
-  max.nterms <- ifnot.null(max.nterms, if (type == "heatmap") NULL else 30L)
   if (missing(theme))
     theme <- getOption("midr.sequential", getOption("midr.qualitative", NULL))
   theme <- color.theme(theme)
@@ -56,7 +55,8 @@ ggmid.mid.importance <- function(
   imp <- object$importance
   if (!is.null(terms))
     imp <- imp[match(terms, imp$term, nomatch = 0L), ]
-  imp <- imp[1L:min(max.nterms, nrow(imp), na.rm = TRUE), ]
+  if (type != "heatmap")
+    imp <- imp[1L:min(max.nterms, nrow(imp), na.rm = TRUE), ]
   # barplot and dotchart
   if (type == "barplot" || type == "dotchart") {
     pl <- ggplot2::ggplot(