+++ R/mid_breakdown.R
@@ -22,7 +22,6 @@
 #' @param data a data frame containing one or more observations for which to calculate the MID breakdown. If not provided, data is automatically extracted based on the function call.
 #' @param row an optional numeric value or character string specifying the row of \code{data} to be used for the breakdown. If \code{NULL}, and the \code{data} contains two or more observations, only the first observation is used.
 #' @param sort logical. If \code{TRUE}, the output data frame is sorted by the absolute contribution of each effect.
-#' @param format an optional named list/vector of character strings. The names should correspond to term names in the model, and the values are format strings passed to \code{sprintf()} to format the feature values (e.g., \code{x1 = "\%.2f"}). If a term is not specified, \code{"\%s"} is used by default for main effects and \code{"\%s, \%s"} for interactions.
 #'
 #' @examples
 #' data(airquality, package = "datasets")
@@ -47,7 +46,8 @@
 #' @export mid.breakdown
 #'
 mid.breakdown <- function(
-    object, data = NULL, row = NULL, sort = TRUE, format = list()) {
+    object, data = NULL, row = NULL, sort = TRUE
+  ) {
   if (is.null(data))
     data <- model.data(object, env = parent.frame())
   if (!is.data.frame(data))
@@ -56,39 +56,19 @@ mid.breakdown <- function(
     data <- data[row, ]
   if (nrow(data) != 1L) {
     message("'data' contains multiple observations: the first observation is used")
-    data <- data[1L, ]
+    data <- data[1L, , drop = FALSE]
   }
   if (nrow(data) == 0L)
     stop("'data' contains no observations to be used")
-  preds <- predict.mid(object, data,
-                       type = "terms", na.action = "na.pass")[1L, ]
-  data <- model.reframe(object, data)
+  preds <- predict.mid(object, data, type = "terms", na.action = "na.pass")[1L, ]
   terms <- names(if (sort) base::sort(abs(preds), decreasing = TRUE) else preds)
-  m <- length(terms)
   orders <- as.factor(sapply(strsplit(terms, split = ":"), length))
-  values <- character(m)
-  if (!is.list(format))
-    format <- as.list(format)
-  for (i in seq_len(m)) {
-    term <- terms[i]
-    tags <- term.split(term)
-    if (length(tags) == 1L) {
-      fmt <- ifnot.null(format[[term]], "%s")
-      u <- data[1L, tags[1L]]
-      values[i] <- sprintf(fmt, u)
-    } else {
-      fmt <- ifnot.null(format[[term]], "%s, %s")
-      u <- data[1L, tags[1L]]
-      v <- data[1L, tags[2L]]
-      values[i] <- sprintf(fmt, u, v)
-    }
-  }
   df <- data.frame(term = factor(terms, levels = rev(terms)),
-                   value = values, mid = preds[terms], order = orders)
+                   mid = preds[terms], order = orders)
   rownames(df) <- NULL
   out <- list()
   out$breakdown <- df
-  out$data <- data
+  out$data <- model.reframe(object, data)
   out$intercept <- object$intercept
   if (!is.null(object$link)) {
     out$linear.predictor <- object$intercept + sum(preds)