+++ NEWS.md
@@ -1,3 +1,9 @@
+# midr 0.5.2.906
+
+-   The `format` argument in `mid.breakdown()` is deprecated.
+-   `plot.mid.breakdown()` and `ggmid.mid.breakdown()` now have a new argument `format.args`, which is passed to `base::format()` to format the predictor values stored in "mid.breakdown" objects.
+-   The `format` argument in `plot.mid.breakdown()` and `ggmid.mid.breakdown()` is renamed to `label.format`. The formatting strings now support more flexible formats, such as "%t=%v, %t=%v" for interactions.
+
 # midr 0.5.2.905
 
 -   Updated `ggmid.mid.importance()` and `plot.mid.importance()` to modify appearance of the plots when color themes are applied.